+++ event-store-management/src/main/java/uk/gov/justice/services/eventstore/management/validation/process/Verifier.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.eventstore.management.validation.process;
+
+import java.util.List;
+
+public interface Verifier {
+
+    List<VerificationResult> verify();
+}