+++ event-store-management/src/test/java/uk/gov/justice/services/eventstore/management/validation/commands/VerifyCatchupCommandHandlerTest.java
@@ -1,0 +1,38 @@
+package uk.gov.justice.services.eventstore.management.validation.commands;
+
+import static org.mockito.Mockito.inOrder;
+
+import uk.gov.justice.services.eventstore.management.validation.process.CatchupVerificationProcess;
+import uk.gov.justice.services.jmx.api.command.VerifyCatchupCommand;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InOrder;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
+import org.slf4j.Logger;
+
+@RunWith(MockitoJUnitRunner.class)
+public class VerifyCatchupCommandHandlerTest {
+
+    @Mock
+    private CatchupVerificationProcess catchupVerificationProcess;
+
+    @Mock
+    private Logger logger;
+
+    @InjectMocks
+    private VerifyCatchupCommandHandler verifyCatchupCommandHandler;
+
+    @Test
+    public void shouldRunTheVerificationProcess() throws Exception {
+
+        verifyCatchupCommandHandler.validateCatchup(new VerifyCatchupCommand());
+
+        final InOrder inOrder = inOrder(logger, catchupVerificationProcess);
+
+        inOrder.verify(logger).info("Received VERIFY_CATCHUP command");
+        inOrder.verify(catchupVerificationProcess).runVerification();
+    }
+}